chore: fix fast refresh
